moved ServiceLog to DTOs
