+++ app/Actions/Server/ClearServiceLog.php
@@ -2,8 +2,8 @@
 
 namespace App\Actions\Server;
 
+use App\DTOs\ServiceLog;
 use App\Models\Server;
-use App\Services\ServiceLog;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Validation\ValidationException;
 use Throwable;
