+++ app/Http/Controllers/ServerLogController.php
@@ -8,12 +8,12 @@ use App\Actions\Server\GetServiceLogs;
 use App\Actions\Server\ReadServiceLog;
 use App\Actions\ServerLog\CreateLog;
 use App\Actions\ServerLog\UpdateLog;
+use App\DTOs\ServiceLog;
 use App\Helpers\QueryBuilder;
 use App\Http\Resources\ServerLogResource;
 use App\Models\Server;
 use App\Models\ServerLog;
 use App\Models\Site;
-use App\Services\ServiceLog;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
