+++ app/Services/HasLogs.php
@@ -2,6 +2,8 @@
 
 namespace App\Services;
 
+use App\DTOs\ServiceLog;
+
 interface HasLogs
 {
     /**
