+++ app/Services/Redis/Redis.php
@@ -2,11 +2,11 @@
 
 namespace App\Services\Redis;
 
+use App\DTOs\ServiceLog;
 use App\Exceptions\ServiceInstallationFailed;
 use App\Exceptions\SSHError;
 use App\Services\AbstractService;
 use App\Services\HasLogs;
-use App\Services\ServiceLog;
 use Closure;
 
 class Redis extends AbstractService implements HasLogs
