+++ app/Services/Webserver/Caddy.php
@@ -3,13 +3,13 @@
 namespace App\Services\Webserver;
 
 use App\Actions\Webserver\GenerateCaddyConfig;
+use App\DTOs\ServiceLog;
 use App\Enums\SslMethod;
 use App\Exceptions\SSHError;
 use App\Exceptions\SSLCreationException;
 use App\Models\Site;
 use App\Models\Ssl;
 use App\Services\HasLogs;
-use App\Services\ServiceLog;
 use Throwable;
 
 class Caddy extends AbstractWebserver implements HasLogs
