+++ app/Services/Webserver/Nginx.php
@@ -3,12 +3,12 @@
 namespace App\Services\Webserver;
 
 use App\Actions\Webserver\GenerateNginxConfig;
+use App\DTOs\ServiceLog;
 use App\Exceptions\SSHError;
 use App\Exceptions\SSLCreationException;
 use App\Models\Site;
 use App\Models\Ssl;
 use App\Services\HasLogs;
-use App\Services\ServiceLog;
 use Throwable;
 
 class Nginx extends AbstractWebserver implements HasLogs
